+++ CHANGELOG.rst
@@ -9,6 +9,8 @@ Changelog
 This version dropped every Python version older than 3.7.
 Please use 2.3 if you need to run ``doc2dash`` on legacy Python versions.
 
+- ``zope.interface`` is not a dependency anymore.
+
 
 ----